+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## [v0.9.5](https://github.com/chipmuenk/pyfda/tree/v0.9.5b1) (2024-12-22)
+## [v0.9.6](https://github.com/chipmuenk/pyfda/tree/v0.9.6) (2025-xx)
+
+### Bugfixes
+
+- Add styling for right-mouse context menu
+- Add missing background color for QFileDialog
+- Fix shrinking of input specification subwidgets instead of turning on scroll bars
+  by setting QSizePolicy() on input_specs widget
+- Fix scaling of loaded data in y[n] tab, int2float and scale_to buttons were disfunct
+  (regression error)
+
+## [v0.9.5](https://github.com/chipmuenk/pyfda/tree/v0.9.5) (2024-12-22)
 
 ### Bugfixes